[#11] chg: nav(header) as sticky
