+++ src/components/layout/index.ts
@@ -1,2 +1,2 @@
-export { default as Nav } from "./nav"
+export { default as Header } from "./header"
 export { default as Footer } from "./footer"
